requirements.txt
fastapi>=0.103.2
uvicorn>=0.23.2
python-multipart>=0.0.6
chromadb>=0.4.14
google-generativeai>=0.3.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
PyMuPDF>=1.23.4
python-docx>=1.0.1
python-dotenv>=1.0.0
